+++ examples/document-search/chroma.py
@@ -36,7 +36,7 @@ import asyncio
 from chromadb import EphemeralClient
 
 from ragbits.core import audit
-from ragbits.core.embeddings.litellm import LiteLLMEmbedder, LiteLLMEmbedderOptions
+from ragbits.core.embeddings.dense import LiteLLMEmbedder, LiteLLMEmbedderOptions
 from ragbits.core.vector_stores.base import VectorStoreOptions
 from ragbits.core.vector_stores.chroma import ChromaVectorStore
 from ragbits.document_search import DocumentSearch, SearchConfig