+++ examples/evaluation/document-search/basic/optimize.py
@@ -46,7 +46,7 @@ config = {
                             },
                             "index_name": "baseline",
                             "embedder": {
-                                "type": "ragbits.core.embeddings.litellm:LiteLLMEmbedder",
+                                "type": "ragbits.core.embeddings.dense:LiteLLMEmbedder",
                                 "config": {
                                     "model_name": "text-embedding-3-small",
                                     "default_options": {