+++ packages/ragbits-core/src/ragbits/core/embeddings/__init__.py
@@ -1,15 +1,16 @@
-from .base import Embedder, EmbedderOptionsT
-from .litellm import LiteLLMEmbedder
-from .noop import NoopEmbedder
-from .sparse import BagOfTokens, SparseEmbedder, SparseEmbedderOptionsT
+from .base import Embedder, EmbedderOptionsT, SparseVector
+from .dense import DenseEmbedder, LiteLLMEmbedder, NoopEmbedder
+from .sparse import BagOfTokens, BagOfTokensOptions, SparseEmbedder, SparseEmbedderOptionsT
 
 __all__ = [
     "BagOfTokens",
-    "BagOfTokens",
+    "BagOfTokensOptions",
+    "DenseEmbedder",
     "Embedder",
     "EmbedderOptionsT",
     "LiteLLMEmbedder",
     "NoopEmbedder",
     "SparseEmbedder",
     "SparseEmbedderOptionsT",
+    "SparseVector",
 ]